Put About Me & Picture into separate file
