+++ src/components/SectionHome/HomeAboutMe.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import Grid from '@material-ui/core/Grid';
+
+const HomePicture = () => {
+  return (
+    <Grid item>
+      <p>About Me</p>
+    </Grid>
+  )
+}
+
+export default HomePicture;
